# api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package api

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	strictnethttp "github.com/oapi-codegen/runtime/strictmiddleware/nethttp"
)

// Defines values for GCPEventAction.
const (
	Create GCPEventAction = "create"
	Delete GCPEventAction = "delete"
	Update GCPEventAction = "update"
)

// Defines values for KubernetesEventType.
const (
	KubernetesEventTypeNormal  KubernetesEventType = "Normal"
	KubernetesEventTypeWarning KubernetesEventType = "Warning"
)

// Defines values for ListKubernetesEventsParamsType.
const (
	ListKubernetesEventsParamsTypeNormal  ListKubernetesEventsParamsType = "Normal"
	ListKubernetesEventsParamsTypeWarning ListKubernetesEventsParamsType = "Warning"
)

// Error defines model for Error.
type Error struct {
	Code    string `json:"code"`
	Message string `json:"message"`
}

// GCPEvent defines model for GCPEvent.
type GCPEvent struct {
	Action       GCPEventAction          `json:"action"`
	Metadata     *map[string]interface{} `json:"metadata,omitempty"`
	ProjectId    string                  `json:"project_id"`
	ResourceName string                  `json:"resource_name"`
	ResourceType string                  `json:"resource_type"`
	Service      string                  `json:"service"`
	Timestamp    time.Time               `json:"timestamp"`
}

// GCPEventAction defines model for GCPEvent.Action.
type GCPEventAction string

// KubernetesEvent defines model for KubernetesEvent.
type KubernetesEvent struct {
	// Count Number of times this event has occurred
	Count int `json:"count"`

	// EventTime ISO-8601 timestamp
	EventTime time.Time `json:"eventTime"`

	// FirstTimestamp ISO-8601 timestamp
	FirstTimestamp time.Time `json:"firstTimestamp"`
	InvolvedObject struct {
		// ApiVersion API version
		ApiVersion string `json:"apiVersion"`

		// FieldPath Field path (if applicable)
		FieldPath *string `json:"fieldPath,omitempty"`

		// Kind Pod|Deployment|Service|etc
		Kind string `json:"kind"`

		// Name Object name
		Name string `json:"name"`

		// Namespace Object namespace
		Namespace string `json:"namespace"`

		// ResourceVersion Resource version
		ResourceVersion *string `json:"resourceVersion,omitempty"`

		// Uid Object UID
		Uid string `json:"uid"`
	} `json:"involvedObject"`

	// LastTimestamp ISO-8601 timestamp
	LastTimestamp time.Time `json:"lastTimestamp"`

	// Message Detailed message about the event
	Message  string `json:"message"`
	Metadata struct {
		Annotations *map[string]interface{} `json:"annotations,omitempty"`

		// CreationTimestamp ISO-8601 timestamp
		CreationTimestamp time.Time               `json:"creationTimestamp"`
		Labels            *map[string]interface{} `json:"labels,omitempty"`

		// Name Event name
		Name string `json:"name"`

		// Namespace Namespace where event occurred
		Namespace string `json:"namespace"`
	} `json:"metadata"`

	// Reason String describing the reason for the event
	Reason string `json:"reason"`
	Source struct {
		// Component Component that generated the event
		Component string `json:"component"`

		// Host Node where event occurred
		Host string `json:"host"`
	} `json:"source"`
	Type KubernetesEventType `json:"type"`
}

// KubernetesEventType defines model for KubernetesEvent.Type.
type KubernetesEventType string

// ListGCPEventsParams defines parameters for ListGCPEvents.
type ListGCPEventsParams struct {
	// ProjectId Filter events by GCP project
	ProjectId *string `form:"project_id,omitempty" json:"project_id,omitempty"`

	// Service Filter events by GCP service
	Service *string `form:"service,omitempty" json:"service,omitempty"`
}

// ListKubernetesEventsParams defines parameters for ListKubernetesEvents.
type ListKubernetesEventsParams struct {
	// Namespace Filter events by namespace
	Namespace *string `form:"namespace,omitempty" json:"namespace,omitempty"`

	// Type Filter events by type
	Type *ListKubernetesEventsParamsType `form:"type,omitempty" json:"type,omitempty"`

	// InvolvedObjectKind Filter events by involved object kind
	InvolvedObjectKind *string `form:"involvedObject.kind,omitempty" json:"involvedObject.kind,omitempty"`
}

// ListKubernetesEventsParamsType defines parameters for ListKubernetesEvents.
type ListKubernetesEventsParamsType string

// CreateGCPEventJSONRequestBody defines body for CreateGCPEvent for application/json ContentType.
type CreateGCPEventJSONRequestBody = GCPEvent

// CreateKubernetesEventJSONRequestBody defines body for CreateKubernetesEvent for application/json ContentType.
type CreateKubernetesEventJSONRequestBody = KubernetesEvent

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// List GCP change events
	// (GET /events/cloud/gcp)
	ListGCPEvents(w http.ResponseWriter, r *http.Request, params ListGCPEventsParams)
	// Create GCP change event
	// (POST /events/cloud/gcp)
	CreateGCPEvent(w http.ResponseWriter, r *http.Request)
	// List Kubernetes change events
	// (GET /events/kubernetes)
	ListKubernetesEvents(w http.ResponseWriter, r *http.Request, params ListKubernetesEventsParams)
	// Create Kubernetes change event
	// (POST /events/kubernetes)
	CreateKubernetesEvent(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// List GCP change events
// (GET /events/cloud/gcp)
func (_ Unimplemented) ListGCPEvents(w http.ResponseWriter, r *http.Request, params ListGCPEventsParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Create GCP change event
// (POST /events/cloud/gcp)
func (_ Unimplemented) CreateGCPEvent(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// List Kubernetes change events
// (GET /events/kubernetes)
func (_ Unimplemented) ListKubernetesEvents(w http.ResponseWriter, r *http.Request, params ListKubernetesEventsParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Create Kubernetes change event
// (POST /events/kubernetes)
func (_ Unimplemented) CreateKubernetesEvent(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// ListGCPEvents operation middleware
func (siw *ServerInterfaceWrapper) ListGCPEvents(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params ListGCPEventsParams

	// ------------- Optional query parameter "project_id" -------------

	err = runtime.BindQueryParameter("form", true, false, "project_id", r.URL.Query(), &params.ProjectId)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "project_id", Err: err})
		return
	}

	// ------------- Optional query parameter "service" -------------

	err = runtime.BindQueryParameter("form", true, false, "service", r.URL.Query(), &params.Service)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "service", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListGCPEvents(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// CreateGCPEvent operation middleware
func (siw *ServerInterfaceWrapper) CreateGCPEvent(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateGCPEvent(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// ListKubernetesEvents operation middleware
func (siw *ServerInterfaceWrapper) ListKubernetesEvents(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params ListKubernetesEventsParams

	// ------------- Optional query parameter "namespace" -------------

	err = runtime.BindQueryParameter("form", true, false, "namespace", r.URL.Query(), &params.Namespace)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "namespace", Err: err})
		return
	}

	// ------------- Optional query parameter "type" -------------

	err = runtime.BindQueryParameter("form", true, false, "type", r.URL.Query(), &params.Type)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "type", Err: err})
		return
	}

	// ------------- Optional query parameter "involvedObject.kind" -------------

	err = runtime.BindQueryParameter("form", true, false, "involvedObject.kind", r.URL.Query(), &params.InvolvedObjectKind)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "involvedObject.kind", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListKubernetesEvents(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// CreateKubernetesEvent operation middleware
func (siw *ServerInterfaceWrapper) CreateKubernetesEvent(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateKubernetesEvent(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/events/cloud/gcp", wrapper.ListGCPEvents)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/events/cloud/gcp", wrapper.CreateGCPEvent)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/events/kubernetes", wrapper.ListKubernetesEvents)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/events/kubernetes", wrapper.CreateKubernetesEvent)
	})

	return r
}

type ListGCPEventsRequestObject struct {
	Params ListGCPEventsParams
}

type ListGCPEventsResponseObject interface {
	VisitListGCPEventsResponse(w http.ResponseWriter) error
}

type ListGCPEvents200JSONResponse []GCPEvent

func (response ListGCPEvents200JSONResponse) VisitListGCPEventsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type ListGCPEvents400JSONResponse Error

func (response ListGCPEvents400JSONResponse) VisitListGCPEventsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type ListGCPEvents500JSONResponse Error

func (response ListGCPEvents500JSONResponse) VisitListGCPEventsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type CreateGCPEventRequestObject struct {
	Body *CreateGCPEventJSONRequestBody
}

type CreateGCPEventResponseObject interface {
	VisitCreateGCPEventResponse(w http.ResponseWriter) error
}

type CreateGCPEvent201JSONResponse GCPEvent

func (response CreateGCPEvent201JSONResponse) VisitCreateGCPEventResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type CreateGCPEvent400JSONResponse Error

func (response CreateGCPEvent400JSONResponse) VisitCreateGCPEventResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type CreateGCPEvent500JSONResponse Error

func (response CreateGCPEvent500JSONResponse) VisitCreateGCPEventResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type ListKubernetesEventsRequestObject struct {
	Params ListKubernetesEventsParams
}

type ListKubernetesEventsResponseObject interface {
	VisitListKubernetesEventsResponse(w http.ResponseWriter) error
}

type ListKubernetesEvents200JSONResponse []KubernetesEvent

func (response ListKubernetesEvents200JSONResponse) VisitListKubernetesEventsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type ListKubernetesEvents400JSONResponse Error

func (response ListKubernetesEvents400JSONResponse) VisitListKubernetesEventsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type ListKubernetesEvents500JSONResponse Error

func (response ListKubernetesEvents500JSONResponse) VisitListKubernetesEventsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type CreateKubernetesEventRequestObject struct {
	Body *CreateKubernetesEventJSONRequestBody
}

type CreateKubernetesEventResponseObject interface {
	VisitCreateKubernetesEventResponse(w http.ResponseWriter) error
}

type CreateKubernetesEvent201JSONResponse KubernetesEvent

func (response CreateKubernetesEvent201JSONResponse) VisitCreateKubernetesEventResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type CreateKubernetesEvent400JSONResponse Error

func (response CreateKubernetesEvent400JSONResponse) VisitCreateKubernetesEventResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type CreateKubernetesEvent500JSONResponse Error

func (response CreateKubernetesEvent500JSONResponse) VisitCreateKubernetesEventResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// List GCP change events
	// (GET /events/cloud/gcp)
	ListGCPEvents(ctx context.Context, request ListGCPEventsRequestObject) (ListGCPEventsResponseObject, error)
	// Create GCP change event
	// (POST /events/cloud/gcp)
	CreateGCPEvent(ctx context.Context, request CreateGCPEventRequestObject) (CreateGCPEventResponseObject, error)
	// List Kubernetes change events
	// (GET /events/kubernetes)
	ListKubernetesEvents(ctx context.Context, request ListKubernetesEventsRequestObject) (ListKubernetesEventsResponseObject, error)
	// Create Kubernetes change event
	// (POST /events/kubernetes)
	CreateKubernetesEvent(ctx context.Context, request CreateKubernetesEventRequestObject) (CreateKubernetesEventResponseObject, error)
}

type StrictHandlerFunc = strictnethttp.StrictHttpHandlerFunc
type StrictMiddlewareFunc = strictnethttp.StrictHttpMiddlewareFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// ListGCPEvents operation middleware
func (sh *strictHandler) ListGCPEvents(w http.ResponseWriter, r *http.Request, params ListGCPEventsParams) {
	var request ListGCPEventsRequestObject

	request.Params = params

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.ListGCPEvents(ctx, request.(ListGCPEventsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ListGCPEvents")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(ListGCPEventsResponseObject); ok {
		if err := validResponse.VisitListGCPEventsResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// CreateGCPEvent operation middleware
func (sh *strictHandler) CreateGCPEvent(w http.ResponseWriter, r *http.Request) {
	var request CreateGCPEventRequestObject

	var body CreateGCPEventJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.CreateGCPEvent(ctx, request.(CreateGCPEventRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreateGCPEvent")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(CreateGCPEventResponseObject); ok {
		if err := validResponse.VisitCreateGCPEventResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// ListKubernetesEvents operation middleware
func (sh *strictHandler) ListKubernetesEvents(w http.ResponseWriter, r *http.Request, params ListKubernetesEventsParams) {
	var request ListKubernetesEventsRequestObject

	request.Params = params

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.ListKubernetesEvents(ctx, request.(ListKubernetesEventsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ListKubernetesEvents")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(ListKubernetesEventsResponseObject); ok {
		if err := validResponse.VisitListKubernetesEventsResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// CreateKubernetesEvent operation middleware
func (sh *strictHandler) CreateKubernetesEvent(w http.ResponseWriter, r *http.Request) {
	var request CreateKubernetesEventRequestObject

	var body CreateKubernetesEventJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.CreateKubernetesEvent(ctx, request.(CreateKubernetesEventRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreateKubernetesEvent")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(CreateKubernetesEventResponseObject); ok {
		if err := validResponse.VisitCreateKubernetesEventResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}